Changed The Rules Of The Game
Now it is max 5 tries each. The player with more points after each
player has tried 5 times, wins. If the score is equal and each player is
out of tries. There will be sudden death. Ive had trouble coding the
logic of the sudden death, so it is in comments.